Add video management and update angularjs to version 1.3.15
Video management includes the possibility to remove, publish and unpublish videos previously uploaded
to the video platform by the watcher. Also add the possibility to edit the video title and description.
And finally add custom properties to videos other than title and description. A dedicated page is available to create new custom properties.