Start watcher from plugin instead of main file
Due to new architecture where the Web Service server is dissociated from the Application server, the watcher needs to be started only on Application server. To to so a new method "start" has been added to all plugins to be able to do something when the plugin is fully loaded on the Application server.