Use new method "process.requireModule" to load modules from openveo-core
Modules not present in openveo-publish, such as winston or express, were not correctly loaded. As NodeJs "require" function looks up in parent directories, if a symbolic link is used, parent node_modules directory won't be in the path. To avoid this problem, modules shared between openveo-core and plugins are loaded with the Function "process.requireModule" available in all plugins.