Transform the package publication process into a final machine state and add mp4 implementation
It wasn't possible to retry a publication when it failed, the only possibility was to remove the package. It is now possible to retry a publication if it fails at some state. It will retry from that state and not from the beginning of the process.
Also publication process can be partially done, if no video platform is specified for a package, the video won't be uploaded automatically. Since then the upload can be launched by the user at any time.
Finally adds support for mp4 package.