Add all available video properties to the VideoModel.add method
Some media videos weren't listed in the add method, consequently those properties could be added to an existing video but not when adding a new one. It is now possible to add a new video specifying the following properties :
available, title, description, properties, originalPackagePath, mediaId, timecodes, chapters, cut and files.