Correct availability bug on a video hosted by Vimeo
Vimeo sets the video as "available" as soon as any definition has been transcoded, not when all definitions have been transcoded. Consequently we have to wait for the expected defintion to be transcoded to set the video as available. Unfortunately the expected video definition can be greater than the one transcoded by Vimeo, in case of non standard video formats. In those cases, the video was blocked forever in non available state. To avoid most of those cases an error factor of 10 has been added.