Implement init method and change start method due to openveo-core modifications
openveo-core now calls the Plugin.init method of each plugin when database and plugins are available, it offers the possibility to create database collections' indexes. Indexes are now created for both properties and videos' collections when the application start.
The Plugin.start method was synchronous, it is now asynchronous.
The Plugin.start method was only called for the application, not the Web Service process, it is now called for each process. Difference between Web Service and application processes can be checked using the process.isWebService property.