Adapt publish to the new version of core and API, with content entities
Entities did not have any relation with a user, thus all users could manipulate all entities. Each content entities has now a specific user (the owner).
A user can be associated to the content entity model which will be used to set the owner of a content entity (add operation) and will be used to test if user has necessary permissions to manipulate a content entity in case of get/remove/update. Super administrator has all power and can get/remove/update any content entity belonging to all users. If no user is specified while adding a content entity, it belongs to the anonymous user.
Each video is now a content entity and have a owner.