Increase error factor to detect Vimeo transcoding end
Vimeo set the video as "available" as soon as any definition has been
transcoded not when all definitions have been transcoded. To deal with
this problem, we wait for the transcoded format corresponding to the
uploaded video. As Vimeo sometimes transcode the original format into
another one we added an error factor of 10 on the video height. It appears
that sometimes an error of 20 lines is found. Error factor is now 64.