Replace chokidar by a custom directories watcher
Using chokidar, it appeared that, sometimes, a resource wasn't detected
at all. chokidar has been removed to use a custom watcher implementation
which intends to correct chokidar issues for OpenVeo Publish purpose.
The new watcher is capable of watching a list of directories
recursively, detect resources at initialization and wait for a file to
be fully copied.