Improve tags file upload errors
An OpenVeo server error on the uploaded file was returning an HTTP code
200 with a translation id as data. It now returns an OpenVeo error (just
like the other requests) with an error code and an HTTP error code 500.
Also the error message was displayed in client side beneath the
file input field. It now only displays error messages in an alert even
errors coming from the front-end server.