Remove deleted users from videos watcher
When a user was deleted his videos were still associated to him. Same thing for the watcher configuration, if the deleted user was the default user to attribute videos to, he still was associated in the database.
When a user is removed, all his videos are now anonymized (set to the anonymous user) and if it's the default user of watcher configuration the association in database is also removed.