Fix watcher detection issue with slow connections
While experimenting difficulties (slowing down connection), during the upload of a package, on the watcher hot folder, the file could be detected when still a .part file. The .part file is a valid package (just not finished yet). Consequently the watcher was considering the file as valid and could not terminate the publication process as the .part file is temporary and was removed before the end of the process.
It now validates that the file format is the right one and that the extension is as expected.