Replace input[type="time"] by ovp-time directive
The time input format depends of the localization
of the browser. For example english format use 12-
hour clock whereas french is in 24-hour clock for-
mat. But the controller wasn't able to parse va-
rious time format.

The ovp-time directive uses three text inputs (for
hours, minutes and seconds) and converts the value
in milliseconds.