Handle points of interest in milliseconds
Points of interest (chapters, tags & cut) units
were stored in percent (of the video duration).

Now the editor handles points of interest in mil-
liseconds, and the server provide a web service to
convert units. This web service must be called by
the player because it can provide the video dura-
tion.