Add thumbnail field to the video edit form
The user can now customize the video thumbnail in
the edit video form.

The picture must be in JPG format, because it will
be renamed to thumbnail.jpg on the server storage.
Actually the thumbnail dimensions are not checked.

To edit the thumbnail of a video, this video must
be uploaded to a provider (because the thumbnail
can't be downloaded before).