Listen to the player needPoiConversion event
The (front) player and (back) editor controllers
are now listening to the OpenVeo Player's needPoi-
Conversion event. This event is triggered when the
points of interest units of a video are in percent
instead of milliseconds.

So when a controller receive this event the con-
version can be launched using the updatePoiAction
of the (backend) VideoController.