Remove deactivation of thumbnail field
In a previous version of the core an
association existed between labels and
fields in forms which led to unexpected
behaviours such as opening the file browser when
clicking on the label of the thumbnail field while
in readble mode. As their is no longer an
association between label and fields, disabling /
enabling the thumbnail field when switching
between editable / readable form is no longer
required.