Add support for TLS
Add support for TLS as a media platform just like
Wowza, Vimeo, local and Youtube.
Medias are not hosted in TLS but in a NFS
directory shared between TLS and OpenVeo.
Consequently it is OpenVeo which copy the media
file on the file system and remove it when
requested. Also the media id is generated by
OpenVeo.