Rename configure PUBLISH state into SYNCHRONIZE
"CONFIGURE" state, in the state machine, was
calling the "configure" method of the media
platform provider with the media resource ids
as argument. Actually no media platform
provider implements this method. The "configure"
method has been renamed into "update" and is now
called with the whole media package and the datas
to update. Also the "CONFIGURE" state has been
renamed into "SYNCHRONIZE".