Fix medias blocked in "pending" state
Uploading more than 3 medias at a time (next
medias are put in the queue) was causing next
medias to be blocked in pending state with the
message "Media package type is not valid". Media
file validation was made when publishing a media
from the queue, even though media could be in an
advanced state (uploading, waiting for upload,
error and so one) with no longer media file in the
original directory. Media file validation is now
made before calling the PublishManager, also
medias leaving the queue to be published are
retried. Publishing a media with an original
package path corresponding to an existing one,
still not publish the media.