Fix the owner and group of new medias
Owner of medias added through the catalogue was
the one defined in Publish configuration.
Media owner is now set to the user uploading the
media. Same thing for groups, groups belonging
to medias added through the catalogue, were groups
defined in Publish configuration. Groups
specified when adding a new media are now
respected even if groups are defined in Publish
configuration.