Refine rules on "publish-get-videos" scope
Rules on "publish-get-videos" scope were
authorizing all POST requests on
"/publish/videos/*" to authorize
"/publish/videos/:id/poi/convert" but due to
the parameter it wasn't possible to do it
differently at the time.
As Core now offers the possibility to include
parameters in rules, the rule has been modified
to be more restrictive.