Use sprites of images for video points of interest
Web service endpoints GET /videos and GET
/videos/:id used to return the list of timecodes
with small URL (using "publish-thumb-200" style),
the list of timecodes now uses a sprite object to
define the small image and its position into the
sprite file. A sprite object is characterized by
the URL of the sprite image and the x/y
coordinates of the small image inside the sprite.
A sprite image is pre-generated for all videos
during publish process.

A migration script has been added to generate a
sprite for all existing videos.

Note that style "publish-thumb-200" has not been
removed, it is still possible to resize the image
on the fly.