Fix bug when adding a video with a thumbnail
Adding a video with a thumbnail wasn't working and
wasn't displaying any error (since version 6.0.0).
The thumnail was removed before it had a chance
to be copied plus the package wasn't considered as
on error. It now sets the package state to error
if a thumbnail was uploaded and couldn't be
copied. Additionally the original thumbnail is
now removed after the copy or when the copy failed.