Fix player time bar preview image
The preview image on the OpenVeo Player wasn't
moving regarding the position of the cursor and
was stuck to the left edge of the player.

The player couldn't compute the size of the time
bar because the player wasn't painted. The player
is now painted by default but not visible.