vimeo: fix blocked videos after upload
Some videos with exotic definition were blocked
and couldn't be viewed due to the algorithm used
to determine if the Vimeo transcoding operation
finished. The algorithm was waiting for a
transcoded video with an height +-= to 64px of the
original video height. It now waits for a
transcoded video height with an height +-= to 22%
of the original video height.