Add search in points of interest
Using web service GET /videos is was possible to
look for a particular keyword inside videos titles
and descriptions using "query" parameter. It is
now possible to also look for the keyword inside
tags and chapters of videos using "searchInPois"
parameter. By default "searchInPois" is
deactivated, set it to "1" to extend the search
into titles and descriptions of tags / chapters.

To achieve this, tags and chapters that were
originally stored inside "publish_videos"
collection are now stored in a dedicated
collection "publish_pois". There is no difference
between a tag and a chapter in "publish_pois"
collection. Difference is only made in videos
documents as they still have the "tags" and
"chapters" properties. "tags" and "chapters"
properties of video now only contain the ids of
corresponding points of interest stored in
"publish_pois" collection.