server: improve search on video descriptions
Searching for a text in video description and
descriptions of tags and chapters using MongoDB
search wasn't working very well as descriptions
might have contained HTML and HTML entities.
All descriptions are now stored both in HTML and
without HTML. The new property containing the
description without HTML is "descriptionText".
querySearch index on video collection and points
of interest collection makes now use of
"descriptionText" instead of "description".
Consequently searching for a text contained in a
description won't be annoyed by the HTML.

This concerns the web service GET /videos when
option useSmartSearch is set to 1.