server: add missing migration to script 3.0.0
OpenVeo Publish < 3.0.0 was manipulating "sources"
property of a video as an Object. Starting with
OpenVeo Publish >= 3.0.0 the "sources" property
is considered an Array.
It means that videos created by OpenVeo Publish <
3.0.0 could led to a JavaScript error crashing
OpenVeo when accessing them.

Migrating from an OpenVeo Publish < 3.0.0 to
OpenVeo Publish >= 11.0.0 won't have this problem
anymore as "sources" property is now transformed
in 3.0.0 migration script.