server: refactor multi-sources algorithm
Multi-sources algorithm had a lot of caveats as it
was designed. It was interfering with several
transition in the publish state machine and with
the video being processed when merging with it.
Multi-sources is now handled in one transition
called "group" at the end of the publication
process. A video is grouped in another one if the
video it wants to merge with is in a stable state
like READY or PUBLISHED, if it isn't it waits for
the video to be in one of these states.