server: fix 7.0.0 migration script
7.0.0 migration wasn't updating the lastState and
lastTransition properties of videos when lastState
was "mediaConfigured" and lastTransition was
"configureMedia" leaving unknown transition and
state to OpenVeo which was crashing when trying to
apply corresponding transition. It could happen
with videos on error or in a transitional state
at the time of migration.

This is due to "mediaConfigured" state renamed
into "mediaSynchronized" and associated transition
"configureMedia" renamed into "synchronizeMedia".