server: fix 11.0.0 migration script
11.0.0 migration wasn't updating the lastState and
lastTransition properties of videos when lastState
was "publicDirectoryPrepared" or "grouped" and
lastTransition was "preparePublicDirectory" or
"group" leaving unknown transition and state to
OpenVeo which was crashing when trying to apply
corresponding transition. It could happen with
videos on error or in a transitional state at the
time of migration.

This is due to renamings in transitions and
states. "grouped" state has been renamed into
"merged" and transition "group" has been renamed
into "merge".
"publicDirectoryPrepared" state and
"preparePublicDirectory" transition has been
removed so "publicDirectoryPrepared" state is
migrated to "metadataRetrieved" state and
"preparePublicDirectory" transition is migrated
into "uploadMedia".