server: add tag "category" fallback for tar files
Tags names were described in tar files with
data.tagname and was named "TagN" if data.tagname
wasn't defined, with "N" an incremental number.
It is now possible to have tags with property
data.category as a fallback to data.tagname. In
order of priority:

1. data.tagname
2. data.category
3. "TagN"

If tar file contains both tags with data.tagname
or data.category, and tags without it, tags
without a name will be named incrementally without
counting tags with a name.