server: fix Youtube videos availability
Loading a Youtube video with property "available"
having value false was blocked with message "This
media is not yet available...".
Availabilty wasn't being checked.

"available" property is now check when false and
when video is a Youtube video. Check is not
performed when "available" property is true AND
video is a Youtube video.