server: fix video public page display
Displaying the public page of a video in local or
Wowza provider was working the first time but not
after that.
It was due to the
VideoController.updateMediaWithPlatformInfo which
wasn't waiting for the media to be updated before
calling its callback function. Consequently, in
some cases like VideoController.getEntityAction or
VideoController.getVideoReadyAction, the media
object reference was modified by
VideoController.resolveResourcesUrls before the
VideoController.updateMediaWithPlatformInfo did
performed the media update in the database.
So the media sources URLs were prefixed by the
streaming server base URL in the database
document.
Consequently, on load, the media sources URLs had
twice the base URL and video couldn't be loaded.