server: aggregate multi-sources points of interest
Uploading several packages with the same name but
on different watched directory creates a
multi-sources video.

The multi-sources video was keeping the data
(timecodes and tags) of only one of the packages
with the same name.

It now aggregates all data of all packages with
the same name. So all timecodes and tags from all
packages with be merged inside the mutli-sources
video.