server: add new archive format support
OpenVeo Publish was able to process archive files
(tar or zip files) containing a .session file, a
video file and optionally a list of images.

A new format of archive are now supported, archive
files (tar or zip files) containing a info.json
file and one or several videos files.

The old format containing a .session file is now
named version 1 format while the new format
containing a info.json file is named version 2
format.

Version 1 format supports a single video, a list
of tags and a list of timecodes.
Version 2 format supports several videos and a
list of tags but no timecodes.

A version 2 format archive containing several
videos will led to a multi-sources video.

Note that, actually, version 1 format is not
deprecated for the benefit of version 2.
Version 1 is used to add a single video with
optionally tags and timecodes while version 2 is
used to add a multi-sources video with optionally
tags but no timecode.
Note that it is also possible to add a
multi-sources video using version 1 format by
uploading several archives of the same name
in different watcher sub directories but it won't
work from the catalog of videos.

Also publication logs have been standardized to be
easier to read.