server: fix hanging packages with same name
It wasn't possible to retry multi packages with
the same name. Only one of the packages was added
to pending packages and the others were added to
the queue.
In fact only one package with a particular name
could be added to pending packages and if it
was a package waiting for another one with the
same name packages were stuck in the queue.

Packages with the same name are now added to
pending packages even if maxConcurrentPackage is
set to 1.