server: add support for archive tag description
Both archive formats supported a property to set
the tag name but didn't have a property to set
the tag description.

Property "text" of archive format version 1 is now
supported to set the tag description.
Property "text" of archive format version 2 does
no longer set the tag name but the tag description
and a new property "tagname" is now supported to
set the tag name instead of "text".

So both archive formats now support setting the
name and description of a tag with the same
properties: "tagname" for the name of the tag and
"text" for the description of the tag.