server: fix multi-sources when ran in parallel
Merging several packages with the same name wasn't
working when all packages were added
simultaneously to the watcher.
Several packages could try to lock another package
at the same time because packages processing could
have been executed on the same JavaScript loop or
another one.

To fix this an in-memory list of locked packages
has been added and made available to all packages
instances. So the package is locked in this list
first and then lockedByPackage property is set in
storage.
Package still need to be marked as locked in
storage because the OpenVeo process can be
restarted at anytime.