server: fix adding thumbnail through catalog
Adding a thumbnail through catalog when adding a
media wasn't working anymore since last commits,
the thumbnail was generated instead.

Since a few commits the states machine was
initializing the media with the media returned by
the storage however the original thumbnail isn't
in the storage.

Both media and media as returned by the storage
are now merge to keep all information of original
media and information from the storage.