server: fix points of interest conversion
Points of interest of old OpenVeo version were
expressed in percentage instead of milliseconds
and automatically converted into milliseconds on
newer versions when viewing a media.
The detection wasn't working anymore since version
11.0.0 and thus points of interest weren't
converted to milliseconds anymore.

In version 11.0.0 points of interest have been
externalized in a different collection and is no
longer part of the medias collection but the old
unit detection was made inside VideoProvider which
is no longer aware of the values of chapters and
tags.
Detection is now performed in VideoController.