server: fix multi-sources medias hosted on Vimeo
A multi-sources media with different definitions
for each source wasn't displayed when trying to
read it. With persistent message: "This media is
not yet available on the streaming platform...".

Vimeo API considers a video available when at
least one transcoded version of the video is
available. It could be a small definition version
of the video or the original version (which is
also transcoded).
OpenVeo Publish waits for the original version by
comparing video height with the media source
height computed during publication steps and
stored in database under metadata ->
profile-settings -> video-height which works
fine for a single source media.
When working with a multi-sources media OpenVeo
Publish waited for the original version of each
source video to be transcoded by Vimeo by
comparing video height with the first source
height which didn't work if sources had different
definitions.

All media sources videos heights are now stored in
database and used to compare them with transcoded
versions made by Vimeo.