server: fix catalog search for non-admin users
Searching by keywords in the catalog of medias
wasn't working for non-admin users.
This bug has been introduced in version 11.0.0
when adding search in points of interest.

The keywords search filters was added to the
already defined OR operator thus in balance with
access filters. Consequently medias managed by the
users AND medias found by keywords were displayed.