server: fix detection of fragmented MP4
FFprobe was used to detect if an MP4 is fragmented
or not guessing that the number of frames is not
set when MP4 file is fragmented.
However some MP4 files can be fragmented and have
a defined number of frames.

MP4Box is now used to detect if the MP4 file is
fragmented or not. MP4Box checks the presence of
the mvex box inside moov box to assert that the
MP4 file is fragmented.